% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsmadez-package.R
\docType{package}
\name{bsmadez-package}
\alias{bsmadez}
\alias{bsmadez-package}
\title{bsmadez: Benchmarking functions and plots for WGBS data}
\description{
Processing of benchamrking data, adding statistics and plots based on gold standard calls.
}
\author{
\strong{Maintainer}: Reka Toth \email{r.toth@dkfz-heidelberg.de}

}
\keyword{internal}
