% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_Corrplot.R
\name{epiGB.CorrPlot}
\alias{epiGB.CorrPlot}
\title{Correlation Plots}
\usage{
epiGB.CorrPlot(epiGB.data = NULL, user_input = NULL, pdf=T, pdf_path=NULL,
                      rel = T, meth = T, abs = T, filename="epiGB.correlation")
}
\arguments{
\item{user_input}{Optional user data input, defaults to NULL.}

\item{rel}{Plotting of the relative distance,  defaults to TRUE.}

\item{meth}{Plotting of the methylation values, defaults to TRUE.}

\item{abs}{Plotting of the absolute distance,  defaults to TRUE.}

\item{pdf}{Optional pdf output, defaults to TRUE.}

\item{pdf_path}{Optional input of a path where the pdf is saved. Defaults to NULL, the pdf is then saved
in the current folder.}

\item{filename}{Optional change of the core filename, default name is "epiGB.correlation".}
}
\value{
This function generates a matrix of pairwise (between the respective pipelines) correlation scatter plots
(for absolute and relative distance from a predefined consensus corridor, as well as methylation levels measured
by different pipelines) alongside the respective correlation coefficients (Pearson correlation) and density plots
for each sample supplied by the benchmarking (and optionally user) data. If pdf output is requested the function
returns three pdf files with the filename extensions _absolute, _methylation and _relative, respectively.
}
\description{
This function plots the distance of preprocessed benchmark and optionally submitted user data
from the consensus corridor as a correlation scatter plot with regression line.
Plotting (without pdf output) the implemented benchmarking data. Note that it is not necessary to supply 'epiGB.data'
since it is implemented within the package.
epiGB.CorrPlot(pdf = F)
The function allows for the user to choose all or some of the plots to be displayed (all function calls default to TRUE).
Plotting only correlation of the relative distance from the consensus corridor for each sample:
}
