% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Distance.R
\name{epiGB.dist}
\alias{epiGB.dist}
\title{Distance Calculation}
\usage{
epiGB.dist(df = NULL)
}
\arguments{
\item{df}{The preprocessed and melted benchmark and optionally provided user data.}
}
\value{
Returns with a data frame with calculated absolute and relative distances
}
\description{
The following function is used to calculate the absolute and
relative distance of benchmark and optionally provided user data from the consensus corridor.
}
