% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_Distplot.R
\name{epiGB.DistPlot}
\alias{epiGB.DistPlot}
\title{Distance Plots}
\usage{
epiGB.DistPlot(epiGB.data = epiGB.data, user_input = NULL, by_sample = T, by_region = T, sum = T, abs = T, rel = T,
                      pdf=T, pdf_path=NULL, filename="epiGB.distplot" )
}
\arguments{
\item{user_input}{Optional user data input, defaults to NULL.}

\item{by_sample}{Plots the relative and/or absolute distance by sample as violin plot, defaults to TRUE.}

\item{by_region}{Plots the relative and/or absolute distance by region as scatter plot, defaults to TRUE.}

\item{sum}{Plots the average distance over all samples and regions as scatter plot.}

\item{abs}{Plotts the absolute distance for all distance plots, defaults to TRUE.}

\item{rel}{Plotts the relative distance for all distance plots, defaults to TRUE.}

\item{pdf}{Optional pdf output, defaults to TRUE.}

\item{pdf_path}{Optional input of a path where the pdf is saved. Defaults to NULL, the pdf is then saved
in the current folder.}

\item{filename}{Optional change of the core filename, default name is "epiGB.distplot".}
}
\value{
this function generates a total of six plots (per data set) serving to visualize the distance of methylation measurements
from the consensus corridor. Either relative or absolute distance of each pipeline by region as scatter plot as well
as by sample in a violin plot is displayed. Additionally, the mean of either absolute or relative distance of each
pipeline over all regions is plotted by sample as scatter plot to summarize the accuracy of each pipeline.
}
\description{
This function plots the relative and/or absolute distance of preprocessed benchmark and
optionally submitted user data from the consensus corridor either by sample or by region.
Plotting (without pdf output) the implemented benchmarking data. Note that it is not necessary to supply 'epiGB.data'
since it is implemented within the package.
The function allows for the user to choose all or some of the plots to be displayed (all function calls default to TRUE).
Generating plots for absolute and relative distance solely by sample:
epiGB.DistPlot(by_region = F, sum = F, abs = F, rel = F, pdf = F)
}
