% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_Heatmap.R
\name{epiGB.Heatmap}
\alias{epiGB.Heatmap}
\title{Heatmaps}
\usage{
epiGB.Heatmap(epiGB.data=epiGB.data, user_input = NULL, meth = T, dist = T, abs = T, rel = T,
                     pdf=T, pdf_path=NULL, filename="epiGB.heatmap")
}
\arguments{
\item{meth}{Plots the correlation of methylation values of all pipelines.}

\item{dist}{Plots the correlation of relative and/or absolute distance of all pipelines.}

\item{abs}{Plots the correlation of absolute distance in both the methylation and distance plot,
defaults to TRUE.}

\item{rel}{Plots the correlation of relative distance in both the methylation and distance plot,
defaults to TRUE.}

\item{pdf}{Optional pdf output, defaults to TRUE.}

\item{pdf_path}{Optional input of a path where the pdf is saved. Defaults to NULL,
the pdf is then saved in the current folder.}

\item{filename}{Optional change of the core filename, default name is "epiGB.heatmap".}

\item{user_input}{Optional user data input, defaults to NULL.}
}
\value{
this function returns two heatmaps visualizing methylation levels and their respective absolute
and relative distance from the consensus corridor by sample over all preselected regions. The user decides which
of the two heatmaps (methylation and/or relative distance) should be plotted (all function calls default to TRUE).
}
\description{
This function creates a heatmap based on the methylation or the relative and/or absolute distance
of preprocessed benchmark and optionally submitted user data from the consensus corridor.
Plotting (without pdf output) the implemented benchmarking data. Note that it is not necessary to supply 'epiGB.data'
since it is implemented within the package.
epiGB.Heatmap(pdf = F)
Plotting of benchmarking data exclusively on the methylation heatmap (without pdf output):
}
