% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Melt.R
\name{epiGB.melt}
\alias{epiGB.melt}
\title{Data Melting}
\usage{
epiGB.melt(df = NULL)
}
\arguments{
\item{value}{The preprocessed data from submitted by epiGB.DataStructure.}
}
\description{
The following function is used to melt the preprocessed benchmark and
optionally provided user data to a format that can be used by ggplot.
}
