% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_Methplot.R
\name{epiGB.MethPlot}
\alias{epiGB.MethPlot}
\title{Methylation Plot}
\usage{
epiGB.MethPlot(epiGB.data = epiGB.data, user_input = NULL,
                      pdf = T, pdf_path = NULL, filename = "epiGB.methplot")
}
\arguments{
\item{pdf}{Optional pdf output, defaults to TRUE.}

\item{pdf_path}{Optional input of a path where the pdf is saved. Defaults to NULL, the pdf is then saved
in the current folder.}

\item{filename}{Optional change of the core filename, default name is "epiGB.methplot".}
}
\value{
This function generates scatter plots visualizing the methylation levels of all pipelines
against all samples, per region. The consensus corridor for each region and sample and its mean is given as an error bar.
}
\description{
This function plots the methylation of preprocessed benchmark and optionally submitted user data
for each region in a scatter plot.
}
\examples{
Plotting (without pdf output) the implemented benchmarking data. Note that it is not necessary to supply 'epiGB.data'
since it is implemented within the package.
epiGB.MethPlot(pdf = F)
}
