% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Preprocessing.R
\name{Preprocess.MethData}
\alias{Preprocess.MethData}
\title{Preprocessing Methrix Data}
\usage{
Preprocess.MethData(benchmarking_data = NULL, user_data = NULL,
  verbose = TRUE)
}
\arguments{
\item{user_data}{Optional user data input in a methrix object. The sample annotation should have the following columns: sample, method, assembly and pipeline.
The samples available are: 5T, 5N, 6T, 6N. The methods can be: WGBS, TWGBS, PBAT or SWIFT. The assembly has to be hg38.}

\item{epiGB.df}{A methrix object containing the bechmarking data.}
}
\value{
A list with datasets processed for plotting. One list element contains the methylation calls and coverage information for one method.
}
\description{
The following function is used to preprocess whole-genome bisulfite sequencing and
tagmented wgbs
benchmarking and optionally submitted user data for subsequent plotting.
}
