% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ranking_Functions.R
\name{rank.byMethod}
\alias{rank.byMethod}
\title{function to run rankings based on method}
\usage{
rank.byMethod(ranking_by_sample = NULL, epiGB.data = NULL)
}
\description{
rankings by method ------
}
