% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ranking_Functions.R
\name{rank.summary}
\alias{rank.summary}
\title{function to summarise rankings by sample, type and method and give the best 3 pipelines for each category}
\usage{
rank.summary(ranking_by_sample = NULL, ranking_by_type = NULL,
  ranking_by_method = NULL, epiGB.data = NULL)
}
\description{
summary ranking ------
}
