% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ranking_Masterfile.R
\name{rankings}
\alias{rankings}
\title{Main function to run all rankings and summarise the results}
\usage{
rankings(benchmarking_data = NULL, user_data = NULL, verbose = TRUE,
  wdir = NULL)
}
\arguments{
\item{benchmarking_data}{A methrix object containing the bechmarking data. If not provided, the package's own data will be used.}

\item{user_data}{Optional user data input in a methrix object. The sample annotation should have the following columns: sample, method, assembly and pipeline.
The samples available are: 5T, 5N, 6T, 6N. The methods can be: WGBS, TWGBS, PBAT or SWIFT. The assembly has to be hg38.}

\item{wdir}{The output directory. If not provided, the working directory will be used. If non-existing, it will be created.}
}
\description{
All rankings ------
}
